refactor: seperate reservation, 선점 정보 저장 테이블
