+++ PlayGround/src/main/java/com/villain/play/ground/PlayGround/recruitment/required.md
@@ -41,7 +41,7 @@
 > 파티 참여 요청 승인제 고민해보기
 > 파티원끼리 단체 체팅
 
-- 선점
+- 선점 reservation
   - Entity 파티
   - Entity 멤버
   - Entity 사용자
