+++ PlayGround/src/test/java/com/villain/play/ground/PlayGround/party/PartyJoinTest.java
@@ -2,6 +2,7 @@ package com.villain.play.ground.PlayGround.party;
 
 import static org.mockito.Mockito.when;
 
+import com.villain.play.ground.PlayGround.reservation.Reservation;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -14,7 +15,6 @@ import org.mockito.junit.jupiter.MockitoExtension;
 @ExtendWith(MockitoExtension.class)
 class PartyJoinTest {
 
-  Party party;
   @InjectMocks
   PartyService service;
 
@@ -33,33 +33,36 @@ class PartyJoinTest {
   @Test
   @DisplayName("풀파티 참가 요청 실패")
   void fullPartyTest() {
-    Reservation request = new Reservation.Builder().member("오드").user("아령").party("0").build();
+    Reservation request = new Reservation.Builder().member("오드").user("아령").party(0).build();
     Assertions.assertThrows(IllegalStateException.class, () -> {service.join(request);});
   }
 
   @Test
   @DisplayName("이미 선점한 멤버로 파티 가입")
   void noSeatTest(){
-    Reservation request = new Reservation.Builder().member("오드").user("아령").party("0").build();
-// 1. party 가져온다.
-//    2. party 멤버 선점 여부 확인한다
+    Party party = new Party(0, "sound cloud", "live and fall", "villain", 0, 6, 6);
+    Reservation request = new Reservation.Builder().member("오드").user("아령").party(0).build();
+    when(partyRepository.getPartyById(request.getParty())).thenReturn(party);
+    // 1. party 가져온다.
+//    2. 선점 관련 서비스 연결, 파티 아이디, 선점 멤버가 동일한 레코드가 이미 존재하는지 확인
+//      : party 멤버 선점 여부 확인한다
 //    2-1. 자리 잇으면 -> 3. 파티 가입 승인
-//    2-2. 자리 없으면 -> 3. 파티 가입 거절
+//    2-2. 자리 없으면 -> IllegalStateException -> 3. 파티 가입 거절
   }
 
   @Test
   @DisplayName("선점 멤버 선택 안하고 파티 가입")
   void failTest(){
-    Reservation request = new Reservation.Builder().user("아령").party("0").build();
-    // 1. 입력값 검증(Service 책임: 비어있는 입력값이 있는가)
+    Reservation request = new Reservation.Builder().user("아령").party(0).build();
+    // 1. 입력값 검증(Service 책임: 비어있는 입력값이 있는가) -> IllegalArgumentException
     // 2. (빈 입력값 존재) - error response entity 처리? 에러 메시지: 재시도 요청하기?
   }
 
   @Test
   @DisplayName("선입금 여부 선택 안하고 파티 가입")
   void failTest2(){
-    Reservation request = new Reservation.Builder().member("오드").user("아령").party("0").build();
-    // 1. 입력값 검증(Service 책임: 비어있는 입력값이 있는가)
+    Reservation request = new Reservation.Builder().member("오드").user("아령").party(0).build();
+    // 1. 입력값 검증(Service 책임: 비어있는 입력값이 있는가) -> IllegalArgumentException
     // 2. (빈 입력값 존재) - error response entity 처리? 에러 메시지: 재시도 요청하기?
   }
